+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed pylace's `append_column_metadata` to refer to the correct internal function.
 - Pylace will try to use Python's internal conversion for floats in `value_to_datum` conversion. 
 
+### Added 
+ - Pylace's Engine now implements deepcopy.
+
 
 ## [python-0.3.1] - 2023-08-28